ARM: mx6: preserve Boot ROM stack in SPL
The i.MX6 Boot ROM starts its' stack at 0x091ffb4 and appears
to be a little less than 256 bytes deep when starting U-Boot.

This was determined experimentally on i.MX6DL and i.MX6SL.

Allow this stack to be preserved to simplify returning to
ROM when SPL is loaded as a plugin.

Signed-off-by: Eric Nelson <eric@nelint.com>