ARM: imx: mx6memcal: allow build of combined SPL+U-Boot
Add an imximage config file to construct u-boot.imx with SPL
as a plugin and use it as a straw-man for the mx6memcal
virtual board.

Note that this breaks the build of SPL because only one value
can be listed for IMX_CONFIG.

As it stands, all imx SPL builds use an almost empty .cfg file
from arch/arm/configs/spl_sd.cfg.

This could be hard-coded for use in building SPL if we just
remove the comments from the file (copyright).

Signed-off-by: Eric Nelson <eric@nelint.com>